+++ api/controllers/stats/modlogs.js
@@ -1,0 +1,13 @@
+module.exports = {
+  friendlyName: "Modlogs",
+
+  description: "Get the number of moderation logs.",
+
+  inputs: {},
+
+  exits: {},
+
+  fn: async function (inputs) {
+    return await sails.models.moderation.count();
+  },
+};